updated a student class
